MTL-1767 SLES15 Build Modifications
The following changes were necessary for building the cloud-init RPM for
SLES.

The patch files referenced in packages/suse/cloud-init.spec.j2 are not
included in this change. The custom patch code in brpm was too difficult to get
working in the timeframe I had, these files are available for an older
cloud-init release at
https://build.opensuse.org/package/show/Cloud:Tools/cloud-init

Many of the packages referenced by read-dependencies do not exist in
SLES15, but are also not necessary for building.